pyproject.toml
[project]
name = "tradebot"
version = "1.0"
description = "simple dex bot in telegram"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiogram~=3.16.0",
    "aiogram-i18n~=1.4",
    "aiohttp~=3.11.11",
    "alembic~=1.14.0",
    "asyncpg~=0.30.0",
    "redis~=5.2.1",
    "sqlalchemy~=2.0.36",
    "msgspec~=0.18.6",
    "pydantic~=2.9.2",
    "pydantic-settings[yaml]~=2.7.0",
    "fluent_runtime~=0.4.0",
    "aiogram-contrib~=1.1.3",
    "tonutils>=0.1.8",
    "swapcoffee-tokens>=1.0.4",
    "pytoniq-core>=0.1.41",
]

[project.urls]
repository = "https://github.com/bI4ckb34rd/tradebot"

[tool.uv]
dev-dependencies = [
    "mypy~=1.14.0",
    "ruff~=0.8.4",
    "ftl-extract>=0.5.0",
]

[tool.black]
line-length = 99
exclude = "\\.?venv|\\.?tests"

[tool.ruff]
target-version = "py38"
line-length = 99
lint.select = [
    "C",
    "DTZ",
    "E",
    "F",
    "I",
    "ICN",
    "N",
    "PLC",
    "PLE",
    "Q",
    "T",
    "W",
    "YTT",
]
lint.ignore = ["N805"]
exclude = [
    ".venv",
    ".idea",
]

[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]
exclude = [
    "venv",
    ".venv",
    ".idea",
    ".tests",
]
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true

[[tool.mypy.overrides]]
module = ["app.telegram.handlers.*"]
strict_optional = false
warn_return_any = false
disable_error_code = ["union-attr"]
